Improve NoChecksumException
It didn't print anything useful. Now it dumps the file path, expected
hash, computed hash, and the downloaded file summary.
